+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 - Fix Herbalism table column order and width
 - Npc Generator added
+- Races, Gender, Sexes, and Classes table controllers added
 
 ## 0.3.0
 - Add style to route if on route